pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "atria_datasets"
version = "0.0.0"                                                                                                    
description = "Atria Datasets: A collection of datasets for training and evaluating AI agents in the Atria framework."
readme = "README.md"
requires-python = "==3.11.*"
license = { file = "LICENSE.txt" }
authors = [{ name = "Saifullah", email = "saifullah.saifullah@dfki.de" }]
dependencies = [
    "atria-core",
    "atria-hub",
    "atria-registry",
    "boto3>=1.39.4",
    "datadings",
    "datasets>=4.0.0",
    "deltalake>=1.0.2",
    "docile-benchmark>=0.3.4",
    "editdistance>=0.8.1",
    "filelock>=3.18.0",
    "fire>=0.7.0",
    "gdown>=5.2.0",
    "hydra-core==1.3.2",
    "hydra_zen==0.14.0",
    "lakefs>=0.11.1",
    "lazy-loader>=0.4",
    "msgpack>=1.1.1",
    "pandarallel>=1.6.5",
    "pandas>=2.3.1",
    "pyarrow-hotfix>=0.7",
    "pycocotools>=2.0.10",
    "pyyaml>=6.0.2",
    "ray>=2.47.1",
    "requests>=2.32.4",
    "sacremoses>=0.1.1",
    "scikit-learn>=1.7.0",
    "textdistance>=4.6.3",
    "tqdm>=4.67.1",
    "webdataset>=1.0.2",
    "wids>=0.1.11",
]


[tool.hatch.metadata]
allow-direct-references = true

[project.optional-dependencies]
test = ["coverage", "pytest"]

[tool.coverage.report]
skip_covered = true
show_missing = true
precision = 2
exclude_lines = [
    'def __repr__',
    'pragma= no cover',
    'raise NotImplementedError',
    'if TYPE_CHECKING=',
    'if typing.TYPE_CHECKING=',
    '@overload',
    '@typing.overload',
    '\(Protocol\)=$',
    'typing.assert_never',
    'assert_never',
    'if __name__ == "__main__":',
]

## Linter and formatter
[tool.ruff]
# cover and extend the default config in https=//docs.astral.sh/ruff/configuration/
extend-exclude = ["alembic"]
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long, handled by black
    "B008",   # do not perform function calls in argument defaults
    "W191",   # indentation contains tabs
    "B904",   # Allow raising exceptions without from e, for HTTPException
    "COM819", # Trailing comma prohibited
    "D100",   # Missing docstring in public module(file)
    "D104",   # Missing docstring in public package
    "D203",   # 1 blank line required before class docstring
    "E201",   # Whitespace after '('
    "E202",   # Whitespace before ')'
    "E203",   # Whitespace before '='
    "E221",   # Multiple spaces before operator
    "E241",   # Multiple spaces after ','
    "E251",   # Unexpected spaces around keyword / parameter equals
    "W291",   # Trailing whitespace
    "W293",   # Blank line contains whitespace
]

isort = { combine-as-imports = true,  split-on-trailing-comma = false }

# Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

[tool.ruff.format]
docstring-code-format = true
skip-magic-trailing-comma = true

[project.urls]
"Homepage" = "https://github.com/saifullah3396/atria/"
"Bug Reports" = "https://github.com/saifullah3396/atria/"
"Source" = "https://github.com/saifullah3396/atria/"

[tool.pytest.ini_options]
pythonpath = [".", "src", "tests"]
# Set additional command line options for pytest
# Ref: https://docs.pytest.org/en/stable/reference/reference.html#command-line-flags
addopts = "-rXs --strict-config --strict-markers --tb=short"
xfail_strict = true         # Treat tests that are marked as xfail but pass as test failures
# filterwarnings = ["error"]  # Treat all warnings as errors

[tool.uv]
managed = true

[tool.uv.sources]
mmcv = { url = "https://download.openmmlab.com/mmcv/dist/cu121/torch2.1.0/mmcv-2.1.0-cp311-cp311-manylinux1_x86_64.whl" }
datadings = { git = "https://github.com/saifullah3396/datadings.git" }
atria-registry = { path = "../atria_registry", editable = true }
atria-hub = { path = "../atria_hub" }
atria-core = { path = "../atria_core" }

[dependency-groups]
dev = [
    "coverage>=7.9.2",
    "factory-boy>=3.3.3",
    "mypy>=1.16.1",
    "pytest>=8.4.1",
    "ruff>=0.12.2",
    "ipykernel>=6.28.1",
]
docsets = [
    "pyarrow-hotfix>=0.7",
    "docile-benchmark==0.3.4",
    "editdistance",
    "sacremoses"
]

[tool.mypy]
disable_error_code = ["import-untyped", "import-not-found"]
